test: move to IntegrationTest
