+++ apollo-openapi/src/test/java/com/ctrip/framework/apollo/openapi/client/ApolloOpenApiClientTest.java
@@ -16,22 +16,10 @@
  */
 package com.ctrip.framework.apollo.openapi.client;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-import com.ctrip.framework.apollo.openapi.dto.OpenAppDTO;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import static org.junit.Assert.assertEquals;
+import org.junit.Test;
 
 public class ApolloOpenApiClientTest {
-
-  private final Logger log = LoggerFactory.getLogger(this.getClass());
 
   @Test
   public void testCreate() {
@@ -44,51 +32,11 @@ public class ApolloOpenApiClientTest {
     assertEquals(someToken, client.getToken());
   }
 
-  @Test
+  @Test(expected = IllegalArgumentException.class)
   public void testCreateWithInvalidUrl() {
     String someInvalidUrl = "someInvalidUrl";
     String someToken = "someToken";
 
-    assertThrows(
-        IllegalArgumentException.class,
-        () -> ApolloOpenApiClient.newBuilder().withPortalUrl(someInvalidUrl).withToken(someToken).build()
-    );
+    ApolloOpenApiClient.newBuilder().withPortalUrl(someInvalidUrl).withToken(someToken).build();
   }
-
-  @Test
-  @Disabled("only for integration test")
-  public void testCreateApp() {
-    String someUrl = "http://localhost:8070";
-//    String someToken = "0627b87948c30517157e8b2a9565e473b5a97323a50128f584838ed10559d3fd";
-    String someToken = "9d0a241e9cb2300f302a875b1195340b2b6f56373cf5ca5d006a3f4e1a46b3ef";
-
-    ApolloOpenApiClient client = ApolloOpenApiClient.newBuilder()
-        .withPortalUrl(someUrl)
-        .withToken(someToken)
-        .withReadTimeout(200 * 1000)
-        .withConnectTimeout(200 * 1000)
-        .build();
-
-    final String appId = "openapi-create-app1";
-
-    {
-      OpenAppDTO openAppDTO = new OpenAppDTO();
-      openAppDTO.setName("openapi create app 测试名字");
-      openAppDTO.setAppId(appId);
-      openAppDTO.setOwnerName("user-test-xxx1");
-      openAppDTO.setOwnerEmail("user-test-xxx1@xxx.com");
-      openAppDTO.setOrgId("orgId1");
-      openAppDTO.setOrgName("orgName1");
-      openAppDTO.setAdmins(new HashSet<>(Arrays.asList(
-          "user-test-xxx2", "user3"
-      )));
-      client.createApp(openAppDTO);
-    }
-
-    List<OpenAppDTO> list = client.getAppsByIds(Collections.singletonList(appId));
-    for (OpenAppDTO openAppDTO : list) {
-      log.info("{}", openAppDTO);
-    }
-  }
-
 }
